requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
attrs==23.2.0
certifi==2024.7.4
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
multidict==6.0.5
peewee==3.17.6
pydantic==2.8.2
pydantic_core==2.20.1
pylatexenc==2.10
python-telegram-bot==21.4
sniffio==1.3.1
typing_extensions==4.12.2
yarl==1.9.4
